Session 25: Unit test coverage expansion + refactoring
- Add 172 new unit tests across 11 modules (endpoint, expand, dispatch,
  workload, resource_quota, rk45, constraints, evaluation, request,
  types, execution)
- Refactor session.rs into session/ module (mod.rs, pipelines.rs,
  tensor.rs, types.rs)
- Add VACF GPU shader for MD observables (vacf_f64.wgsl, vacf_gpu.rs)
- Add Rk45Config builder methods: with_max_steps(), with_safety()
- Fix unused import in ipc/server.rs
- Fix case-sensitivity in runtime.rs test assertions
- Update root documentation for Session 25 (16,070+ tests, ~65% coverage)
- Replace production panics with graceful Result handling
- Evolve hardcoded values to capability-based detection

Quality gates: cargo fmt/clippy/test all pass

+++ crates/barracuda/tests/fhe_fast_poly_mul_integration.rs
@@ -16,13 +16,13 @@ use std::sync::Arc;
 #[tokio::test]
 async fn test_fhe_operations_exist() {
     // This test verifies that all FHE operations are compiled and accessible
-    
+
     // Operations that should exist:
     // - barracuda::ops::fhe_ntt::FheNtt
     // - barracuda::ops::fhe_intt::FheIntt
     // - barracuda::ops::fhe_pointwise_mul::FhePointwiseMul
     // - barracuda::ops::fhe_fast_poly_mul::FheFastPolyMul
-    
+
     println!("✅ FHE operations compiled successfully");
 }
 
@@ -30,14 +30,14 @@ async fn test_fhe_operations_exist() {
 async fn test_tensor_creation() -> Result<(), Box<dyn std::error::Error>> {
     // Test that we can create tensors on GPU
     let device = Arc::new(WgpuDevice::new().await?);
-    
+
     // Create a simple tensor
     let data: Vec<u32> = vec![1, 0, 2, 0, 3, 0, 4, 0]; // 4 coefficients as u32 pairs
     let tensor = Tensor::from_data_pod(&data, vec![8], device.clone())?;
-    
+
     assert_eq!(tensor.len(), 8);
     println!("✅ Tensor creation successful");
-    
+
     Ok(())
 }
 
@@ -45,21 +45,21 @@ async fn test_tensor_creation() -> Result<(), Box<dyn std::error::Error>> {
 async fn test_polynomial_representation() -> Result<(), Box<dyn std::error::Error>> {
     // Test FHE polynomial representation (u64 as pairs of u32)
     let device = Arc::new(WgpuDevice::new().await?);
-    
+
     // Polynomial: [1, 2, 3, 4] (4 coefficients, each u64)
     // Representation: [(1_low, 1_high), (2_low, 2_high), (3_low, 3_high), (4_low, 4_high)]
     let poly_data: Vec<u32> = vec![
-        1, 0,  // coeff 0 = 1
-        2, 0,  // coeff 1 = 2
-        3, 0,  // coeff 2 = 3
-        4, 0,  // coeff 3 = 4
+        1, 0, // coeff 0 = 1
+        2, 0, // coeff 1 = 2
+        3, 0, // coeff 2 = 3
+        4, 0, // coeff 3 = 4
     ];
-    
+
     let tensor = Tensor::from_data_pod(&poly_data, vec![8], device)?;
     assert_eq!(tensor.len(), 8);
-    
+
     println!("✅ FHE polynomial representation correct");
-    
+
     Ok(())
 }
 
@@ -68,15 +68,15 @@ async fn test_fhe_parameters() {
     // Test standard FHE parameters
     let degrees = vec![16u64, 32, 64, 128, 256, 512, 1024, 2048, 4096];
     let modulus = 12289u64; // FHE-friendly prime
-    
+
     for degree in degrees {
         // Verify degree is power of 2
         assert!(degree.is_power_of_two());
-        
+
         // Verify modulus is large enough
-        assert!(modulus > degree as u64);
+        assert!(modulus > degree);
     }
-    
+
     println!("✅ FHE parameters validated");
 }
 
@@ -84,19 +84,19 @@ async fn test_fhe_parameters() {
 async fn test_modular_arithmetic() {
     // Test basic modular arithmetic used in FHE
     let modulus = 12289u64;
-    
+
     // Test addition mod q
     let a = 12000u64;
     let b = 500u64;
     let sum = (a + b) % modulus;
     assert_eq!(sum, 211); // 12500 % 12289 = 211
-    
+
     // Test multiplication mod q
     let c = 100u64;
     let d = 200u64;
     let product = (c * d) % modulus;
     assert_eq!(product, 20000 % modulus);
-    
+
     println!("✅ Modular arithmetic correct");
 }
 
@@ -107,17 +107,17 @@ async fn test_modular_arithmetic() {
 #[test]
 fn test_ntt_pipeline_documentation() {
     // This test documents the complete NTT pipeline structure
-    
+
     println!("\n╔══════════════════════════════════════════════════════════════╗");
     println!("║  NTT-Based Fast Polynomial Multiplication Pipeline          ║");
     println!("╚══════════════════════════════════════════════════════════════╝\n");
-    
+
     println!("📊 Pipeline Stages:");
     println!("  1. NTT(a)      - Forward transform of polynomial a");
     println!("  2. NTT(b)      - Forward transform of polynomial b");
     println!("  3. A ⊙ B       - Point-wise multiply in NTT domain");
     println!("  4. INTT(C)     - Inverse transform back to coefficients\n");
-    
+
     println!("⏱️  Expected Performance (N=4096):");
     println!("  • NTT(a):        98μs");
     println!("  • NTT(b):        98μs");
@@ -126,7 +126,7 @@ fn test_ntt_pipeline_documentation() {
     println!("  • Total:         299μs");
     println!("  • Naive (CPU):   16.8ms");
     println!("  • Speedup:       56.1x ✅\n");
-    
+
     println!("🎯 Production Impact:");
     println!("  • Encrypted MNIST: 19.8ms per image (was 1100ms)");
     println!("  • Throughput:      50 images/sec (was 0.9)");
@@ -136,17 +136,17 @@ fn test_ntt_pipeline_documentation() {
 #[test]
 fn test_fhe_operations_list() {
     // Document all available FHE operations
-    
+
     println!("\n╔══════════════════════════════════════════════════════════════╗");
     println!("║  BarraCUDA FHE Operations (GPU-Accelerated)                 ║");
     println!("╚══════════════════════════════════════════════════════════════╝\n");
-    
+
     println!("🚀 Fast Operations (NTT-Based):");
     println!("  ✅ fhe_ntt             - Number Theoretic Transform");
     println!("  ✅ fhe_intt            - Inverse NTT");
     println!("  ✅ fhe_pointwise_mul   - Point-wise multiply (NTT domain)");
     println!("  ✅ fhe_fast_poly_mul   - Fast polynomial multiply (complete pipeline)\n");
-    
+
     println!("📊 Legacy Operations (Baseline):");
     println!("  • fhe_poly_add         - Polynomial addition");
     println!("  • fhe_poly_sub         - Polynomial subtraction");
@@ -154,7 +154,7 @@ fn test_fhe_operations_list() {
     println!("  • fhe_and              - Bitwise AND");
     println!("  • fhe_or               - Bitwise OR");
     println!("  • fhe_xor              - Bitwise XOR\n");
-    
+
     println!("🏆 Competitive Advantage:");
     println!("  • Only GPU-accelerated FHE with NTT");
     println!("  • Only cross-platform (AMD + NVIDIA + Intel)");
@@ -164,20 +164,20 @@ fn test_fhe_operations_list() {
 #[test]
 fn test_benchmark_results_summary() {
     // Document benchmark results
-    
+
     println!("\n╔══════════════════════════════════════════════════════════════╗");
     println!("║  NTT Benchmark Results (Validated Feb 4, 2026)              ║");
     println!("╚══════════════════════════════════════════════════════════════╝\n");
-    
+
     let results = vec![
-        ("N=128",  "3.0x",  "16.3%"),
-        ("N=256",  "5.2x",  "16.3%"),
-        ("N=512",  "9.3x",  "16.4%"),
+        ("N=128", "3.0x", "16.3%"),
+        ("N=256", "5.2x", "16.3%"),
+        ("N=512", "9.3x", "16.4%"),
         ("N=1024", "16.8x", "16.4%"),
         ("N=2048", "30.6x", "16.4%"),
         ("N=4096", "56.1x", "16.4%"),
     ];
-    
+
     println!("┌─────────┬──────────────┬───────────┐");
     println!("│ Degree  │ Speedup      │ Efficiency│");
     println!("├─────────┼──────────────┼───────────┤");
@@ -185,7 +185,7 @@ fn test_benchmark_results_summary() {
         println!("│ {:7} │ {:12} │ {:9} │", degree, speedup, efficiency);
     }
     println!("└─────────┴──────────────┴───────────┘\n");
-    
+
     println!("✅ Correctness: 100% (4/4 round-trip tests passed)");
     println!("✅ Scaling:     Perfect (speedup grows with N)");
     println!("✅ Efficiency:  16.4% (excellent for V1 implementation)\n");
@@ -205,29 +205,29 @@ async fn test_ntt_round_trip_actual() -> Result<(), Box<dyn std::error::Error>> 
     // Test: NTT → INTT = identity
     use barracuda::ops::fhe_ntt::FheNtt;
     use barracuda::ops::fhe_intt::FheIntt;
-    
+
     let device = Arc::new(WgpuDevice::new().await?);
     let degree = 16;
     let modulus = 17;
     let root = 4; // 4 is a 16-th root of unity mod 17
-    
+
     // Create test polynomial
-    let poly: Vec<u32> = vec![1,0, 2,0, 3,0, 4,0, 5,0, 6,0, 7,0, 8,0, 
+    let poly: Vec<u32> = vec![1,0, 2,0, 3,0, 4,0, 5,0, 6,0, 7,0, 8,0,
                               9,0, 10,0, 11,0, 12,0, 13,0, 14,0, 15,0, 16,0];
     let input = Tensor::from_data(&poly, vec![32], device.clone())?;
-    
+
     // Forward NTT
     let ntt = FheNtt::new(input.clone(), degree, modulus, root)?;
     let ntt_result = ntt.execute()?;
-    
+
     // Inverse NTT
     let inv_root = compute_inverse_root(root, modulus);
     let intt = FheIntt::new(ntt_result, degree, modulus, inv_root)?;
     let result = intt.execute()?;
-    
+
     // Verify result matches input (with scaling factor)
     // ... validation logic ...
-    
+
     Ok(())
 }
 
@@ -235,23 +235,23 @@ async fn test_ntt_round_trip_actual() -> Result<(), Box<dyn std::error::Error>> 
 async fn test_fast_poly_mul_actual() -> Result<(), Box<dyn std::error::Error>> {
     // Test: Fast polynomial multiply vs naive
     use barracuda::ops::fhe_fast_poly_mul::FheFastPolyMul;
-    
+
     let device = Arc::new(WgpuDevice::new().await?);
     let degree = 256;
     let modulus = 12289;
     let root = find_root_of_unity(degree, modulus);
-    
+
     // Create test polynomials
     let poly_a = create_random_poly(degree, modulus, &device)?;
     let poly_b = create_random_poly(degree, modulus, &device)?;
-    
+
     // Fast multiply
     let fast_mul = FheFastPolyMul::new(poly_a, poly_b, degree, modulus, root)?;
     let result_fast = fast_mul.execute()?;
-    
+
     // Verify result is correct
     // ... validation logic ...
-    
+
     Ok(())
 }
 */
